+++ src/it/scala/au/com/agl/arc/execute/KafkaCommitExecuteSuite.scala
@@ -11,6 +11,7 @@ import org.apache.commons.io.FileUtils
 import org.apache.commons.io.IOUtils
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.types._
 
 import au.com.agl.arc.api._
 import au.com.agl.arc.api.API._
@@ -90,7 +91,9 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType
       )
     ).get
 
@@ -114,7 +117,9 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType
       )
     ).get
 
@@ -150,7 +155,9 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType
       )
     ).get
     actual = extractDataset2.select($"value").as[String]
@@ -194,7 +201,9 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType
       )
     ).get