+++ src/it/scala/au/com/agl/arc/extract/KafkaExtractSuite.scala
@@ -11,6 +11,7 @@ import org.apache.commons.io.FileUtils
 import org.apache.commons.io.IOUtils
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.types._
 
 import au.com.agl.arc.api._
 import au.com.agl.arc.api.API._
@@ -47,7 +48,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
     FileUtils.deleteQuietly(new java.io.File(checkPointPath)) 
   }
 
-  test("KafkaExtract") {
+  test("KafkaExtract: String") {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
@@ -63,6 +64,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
       .withColumn("normal", randn(seed=27))
       .repartition(10)
       .toJSON
+
     dataset.createOrReplaceTempView(inputView)
 
     load.KafkaLoad.load(
@@ -94,12 +96,85 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType
       )
     ).get
 
     val expected = dataset
-    val actual = extractDataset.select($"value").as[String]
+    val actual = extractDataset.select(col("value")).as[String]
+
+    val actualExceptExpectedCount = actual.except(expected).count
+    val expectedExceptActualCount = expected.except(actual).count
+    if (actualExceptExpectedCount != 0 || expectedExceptActualCount != 0) {
+      println("actual")
+      actual.show(false)
+      println("expected")
+      expected.show(false)  
+    }
+    assert(actual.except(expected).count === 0)
+    assert(expected.except(actual).count === 0)
+  }  
+
+
+  test("KafkaExtract: Binary") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
+
+    val topic = UUID.randomUUID.toString
+    val groupId = UUID.randomUUID.toString
+
+    // insert 100 records
+    val dataset = spark.sqlContext.range(0, 100)
+      .select("id")
+      .withColumn("uniform", rand(seed=10))
+      .withColumn("normal", randn(seed=27))
+      .repartition(10)
+      .toJSON
+      .select(col("value").cast(BinaryType))
+
+    dataset.createOrReplaceTempView(inputView)
+
+    load.KafkaLoad.load(
+      KafkaLoad(
+        name="df", 
+        description=None,
+        inputView=inputView, 
+        topic=topic,
+        bootstrapServers=bootstrapServers,
+        acks= -1,
+        numPartitions=None, 
+        batchSize=16384, 
+        retries=0, 
+        params=Map.empty
+      )
+    )   
+
+    val extractDataset = extract.KafkaExtract.extract(
+      KafkaExtract(
+        name="df", 
+        description=None,
+        outputView=outputView, 
+        topic=topic,
+        bootstrapServers=bootstrapServers,
+        groupID=groupId,
+        maxPollRecords=10000, 
+        timeout=timeout, 
+        autoCommit=false, 
+        persist=true, 
+        numPartitions=None, 
+        partitionBy=Nil,
+        params=Map.empty,
+        keyType=BinaryType,
+        valueType=BinaryType
+      )
+    ).get
+
+    val expected = dataset
+    val actual = extractDataset.select(col("value"))
 
     val actualExceptExpectedCount = actual.except(expected).count
     val expectedExceptActualCount = expected.except(actual).count
@@ -130,6 +205,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
       .repartition(10)
       .toJSON
     dataset.createOrReplaceTempView(inputView)
+
     load.KafkaLoad.load(
       KafkaLoad(
         name="df", 
@@ -159,7 +235,9 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType        
       )
     ).get
 
@@ -177,7 +255,9 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType   
       )
     ).get
 
@@ -215,6 +295,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
       .repartition(10)
       .toJSON
     dataset.createOrReplaceTempView(inputView)
+
     load.KafkaLoad.load(
       KafkaLoad(
         name="df", 
@@ -244,7 +325,9 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType   
       )
     ).get
 
@@ -262,7 +345,9 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType   
       )
     ).get
 
@@ -297,7 +382,9 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty,
+        keyType=StringType,
+        valueType=StringType   
       )
     ).get
 
@@ -320,7 +407,6 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
       .option("kafka.bootstrap.servers", bootstrapServers)
       .option("topic", topic)
       .start()
-
 
     val df = spark.table("extract")